+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/components/reactnativestripesdk/CardFormViewManager.kt
@@ -7,19 +7,19 @@ import com.facebook.react.uimanager.SimpleViewManager
 import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.annotations.ReactProp
 
-class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
-  override fun getName() = "CardField"
+class CardFormViewManager : SimpleViewManager<CardFormView>() {
+  override fun getName() = "CardForm"
 
   private var reactContextRef: ThemedReactContext? = null
 
   override fun getExportedCustomDirectEventTypeConstants(): MutableMap<String, Any> {
     return MapBuilder.of(
       CardFocusEvent.EVENT_NAME, MapBuilder.of("registrationName", "onFocusChange"),
-      CardChangedEvent.EVENT_NAME, MapBuilder.of("registrationName", "onCardChange")
+      CardFormCompleteEvent.EVENT_NAME, MapBuilder.of("registrationName", "onFormComplete")
     )
   }
 
-  override fun receiveCommand(root: StripeSdkCardView, commandId: String?, args: ReadableArray?) {
+  override fun receiveCommand(root: CardFormView, commandId: String?, args: ReadableArray?) {
     when (commandId) {
       "focus" -> root.requestFocusFromJS()
       "blur" -> root.requestBlurFromJS()
@@ -28,45 +28,45 @@ class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
   }
 
   @ReactProp(name = "dangerouslyGetFullCardDetails")
-  fun setDangerouslyGetFullCardDetails(view: StripeSdkCardView, dangerouslyGetFullCardDetails: Boolean = false) {
+  fun setDangerouslyGetFullCardDetails(view: CardFormView, dangerouslyGetFullCardDetails: Boolean = false) {
     view.setDangerouslyGetFullCardDetails(dangerouslyGetFullCardDetails)
   }
 
   @ReactProp(name = "postalCodeEnabled")
-  fun setPostalCodeEnabled(view: StripeSdkCardView, postalCodeEnabled: Boolean = true) {
+  fun setPostalCodeEnabled(view: CardFormView, postalCodeEnabled: Boolean = false) {
     view.setPostalCodeEnabled(postalCodeEnabled)
   }
 
+  // @ReactProp(name = "placeholder")
+  // fun setPlaceHolders(view: CardFormView, placeholder: ReadableMap) {
+  //   view.setPlaceHolders(placeholder);
+  // }
+
   @ReactProp(name = "autofocus")
-  fun setAutofocus(view: StripeSdkCardView, autofocus: Boolean = false) {
+  fun setAutofocus(view: CardFormView, autofocus: Boolean = false) {
     view.setAutofocus(autofocus)
   }
 
   @ReactProp(name = "cardStyle")
-  fun setCardStyle(view: StripeSdkCardView, cardStyle: ReadableMap) {
+  fun setCardStyle(view: CardFormView, cardStyle: ReadableMap) {
     view.setCardStyle(cardStyle)
   }
 
-  @ReactProp(name = "placeholder")
-  fun setPlaceHolders(view: StripeSdkCardView, placeholder: ReadableMap) {
-    view.setPlaceHolders(placeholder)
-  }
-
-  override fun createViewInstance(reactContext: ThemedReactContext): StripeSdkCardView {
+  override fun createViewInstance(reactContext: ThemedReactContext): CardFormView {
     val stripeSdkModule: StripeSdkModule? = reactContext.getNativeModule(StripeSdkModule::class.java)
-    val view = StripeSdkCardView(reactContext)
+    val view = CardFormView(reactContext)
 
     reactContextRef = reactContext
 
-    stripeSdkModule?.cardFieldView = view
+    stripeSdkModule?.cardFormView = view
     return view
   }
 
-  override fun onDropViewInstance(view: StripeSdkCardView) {
+  override fun onDropViewInstance(view: CardFormView) {
     super.onDropViewInstance(view)
 
     val stripeSdkModule: StripeSdkModule? = reactContextRef?.getNativeModule(StripeSdkModule::class.java)
-    stripeSdkModule?.cardFieldView = null
+    stripeSdkModule?.cardFormView = null
     reactContextRef = null
   }
 }